+++ application/miniprogram/components/order/index.json
@@ -1,0 +1,9 @@
+{
+  "component": true,
+  "usingComponents": {
+    "van-panel": "@vant/weapp/panel/index",
+    "van-card": "@vant/weapp/card/index",
+    "van-cell": "@vant/weapp/cell/index",
+    "van-button": "@vant/weapp/button/index"
+  }
+}
