+++ application/miniprogram/pages/order/index.wxml
@@ -1,17 +1,11 @@
-<van-tabs active="{{ active }}" bind:change="onChange">
+<van-tabs active="{{ index }}" bind:change="onChange">
   <van-tab title="全部">
-    <van-panel title="汇光皇后卤肉" desc="{{'订单号：' + item.orderNumber}}" status="{{item.active}}" wx:for="{{order}}" wx:key="id">
-      <van-card
-        wx:for="{{item.goods}}"
-        wx:key="index"
-        num="{{item.count}}"
-        price="{{item.originPrice}}"
-        desc="{{item.desc}}"
-        title="商品标题"
-        thumb="{{ 'https://6465-dev-o45qm-1301624908.tcb.qcloud.la/goods-list/10141584450128_.pic.jpg' }}"
-      />
-    </van-panel>
+    <order bindmyevent="turnPage" id="order"></order>
   </van-tab>
-  <van-tab title="待发货">待发货</van-tab>
-  <van-tab title="已收货">已收货</van-tab>
+  <van-tab title="待发货">
+    <order bindmyevent="turnPage" id="order"></order>
+  </van-tab>
+  <van-tab title="已收货">
+    <order bindmyevent="turnPage" id="order"></order>
+  </van-tab>
 </van-tabs>
